+++ mplotutils/tests/__init__.py
@@ -1,8 +1,17 @@
 import contextlib
+import warnings
 
 import matplotlib
 import matplotlib.pyplot as plt
 import pytest
+
+
+@contextlib.contextmanager
+def assert_no_warnings():
+
+    with warnings.catch_warnings(record=True) as record:
+        yield
+        assert len(record) == 0, "got unexpected warning(s)"
 
 
 @contextlib.contextmanager